+++ src/Sulu/Bundle/WebsiteBundle/EventListener/AppendAnalyticsListener.php
@@ -66,7 +66,11 @@ class AppendAnalyticsListener
         }
 
         $portalUrl = $this->requestAnalyzer->getAttribute('urlExpression');
-        $analytics = $this->analyticsRepository->findByUrl($portalUrl, $this->environment);
+        $analytics = $this->analyticsRepository->findByUrl(
+            $portalUrl,
+            $this->requestAnalyzer->getPortalInformation()->getWebspaceKey(),
+            $this->environment
+        );
 
         $content = $this->engine->render('SuluWebsiteBundle:Analytics:website.html.twig', ['analytics' => $analytics]);